backend/requirements.txt
# requirements.txt (UPDATE)

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyjwt==2.8.0
PyYAML==6.0.1
python-dotenv==1.0.0
cors==1.0.1
paypalrestsdk
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1